# app/Http/Controllers/Api/SalesExecutiveLeadController.php
<?php

namespace App\Http\Controllers\Api;

use App\Http\Controllers\Controller;
use App\Models\Lead;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Auth;
use Illuminate\Support\Facades\Log;
use Carbon\Carbon;
use App\Models\LeadStatus;

class SalesExecutiveLeadController extends Controller
{
    /**
     * @OA\Get(
     *     path="/api/sales-executive/leads",
     *     summary="Get sales executive leads",
     *     description="Retrieve all leads assigned to the logged-in sales executive",
     *     operationId="getSalesExecutiveLeads",
     *     tags={"Leads"},
     *     security={{"sanctum": {}}},
     *     @OA\Response(
     *         response=200,
     *         description="Sales executive leads retrieved successfully",
     *         @OA\JsonContent(
     *             type="array",
     *             @OA\Items(
     *                 @OA\Property(property="id", type="integer", example=1),
     *                 @OA\Property(property="shop_name", type="string", example="ABC Electronics"),
     *                 @OA\Property(property="contact_person", type="string", example="John Doe"),
     *                 @OA\Property(property="mobile_number", type="string", example="9876543210"),
     *                 @OA\Property(property="email", type="string", example="john@example.com"),
     *                 @OA\Property(property="address", type="string", example="123 Main Street, City"),
     *                 @OA\Property(property="area_locality", type="string", example="Downtown"),
     *                 @OA\Property(property="pincode", type="string", example="123456"),
     *                 @OA\Property(property="business_type", type="integer", example=1),
     *                 @OA\Property(property="current_system", type="integer", example=1),
     *                 @OA\Property(property="lead_status", type="integer", example=1),
     *                 @OA\Property(property="plan_interest", type="string", example="Premium Plan"),
     *                 @OA\Property(property="next_follow_up_date", type="string", format="date", example="2024-01-15"),
     *                 @OA\Property(property="meeting_notes", type="string", example="Initial discussion completed"),
     *                 @OA\Property(property="assigned_to", type="integer", example=1),
     *                 @OA\Property(property="created_at", type="string", format="date-time", example="2024-01-01T10:00:00Z"),
     *                 @OA\Property(property="updated_at", type="string", format="date-time", example="2024-01-01T10:00:00Z"),
     *                 @OA\Property(property="lead_status_data", type="object", example={"id": 1, "name": "New"})
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthenticated"
     *     )
     * )
     */
    public function index()
    {
        // get all leads assigned to the logged-in sales executive
        $leads = Lead::where('assigned_to', Auth::id())
            ->with('leadStatusData')
            ->get();
        return response()->json($leads);
    }

    /**
     * @OA\Get(
     *     path="/api/sales-executive/lead-counts-by-status",
     *     summary="Get lead counts by status",
     *     description="Get lead counts grouped by status for the logged-in sales executive",
     *     operationId="getLeadCountsByStatus",
     *     tags={"Leads"},
     *     security={{"sanctum": {}}},
     *     @OA\Response(
     *         response=200,
     *         description="Lead counts by status retrieved successfully",
     *         @OA\JsonContent(
     *             type="array",
     *             @OA\Items(
     *                 @OA\Property(property="lead_status", type="integer", example=1),
     *                 @OA\Property(property="status_name", type="string", example="New"),
     *                 @OA\Property(property="total", type="integer", example=5)
     *             )
     *         )
     *     ),
     *     @OA\Response(
     *         response=401,
     *         description="Unauthenticated"
     *     )
     * )
     */
    public function leadCountsByStatus()
    {
        $userId = Auth::id();
        $counts = Lead::where('assigned_to', $userId)
            ->select('lead_status', \DB::raw('count(*) as total'))
            ->groupBy('lead_status')
            ->with('leadStatusData:id,name') // Eager load lead status details
            ->get()
            ->map(function ($item) {
                return [
                    'lead_status' => $item->lead_status,
                    'status_name' => $item->leadStatusData ? $item->leadStatusData->name : null,
                    'total' => $item->total,
                ];
            });
        return response()->json($counts);
    }

    // get all leads assigned to the logged-in sales executive with pagination, search, and filters
    /* 
    GET /sales-executive/leads?search=sharma&lead_status[]=Open&lead_status[]=Warm&follow_up_start_date=2025-08-01&follow_up_end_date=2025-08-31&sort=next_follow_up_date&direction=asc&per_page=20
    */
    public function leads(Request $request)
    {

        $query = Lead::query()
            ->where('assigned_to', Auth::id())
            ->with('leadStatusData');

        // Keyword search across multiple columns
        if ($search = trim((string) $request->input('search'))) {
            $query->where(function ($q) use ($search) {
                $q->where('shop_name', 'like', "%{$search}%")
                    ->orWhere('contact_person', 'like', "%{$search}%")
                    ->orWhere('mobile_number', 'like', "%{$search}%")
                    ->orWhere('email', 'like', "%{$search}%")
                    ->orWhere('address', 'like', "%{$search}%")
                    ->orWhere('area_locality', 'like', "%{$search}%")
                    ->orWhere('pincode', 'like', "%{$search}%");
            });
        }

        // Get No follow up statuses
        $noFollowUpStatuses = LeadStatus::whereIn('name', ['Sold', 'Already Using CRM', 'Not Interested', 'Using Different App'])->pluck('id');

        // if lead_status is today then return only today's where follow_up_date is todays date
        if ($request->filled('lead_status') && $request->input('lead_status') == 'today') {
            $query->whereDate('next_follow_up_date', '<=', Carbon::today())
            ->whereNotIn('lead_status', $noFollowUpStatuses);
        }        

        // Filter by lead_status (supports single value or array of values)
        if ($request->filled('lead_status') && $request->input('lead_status') != 'today') {
            $statuses = $request->input('lead_status');
            $statuses = is_array($statuses) ? $statuses : [$statuses];
            $query->whereIn('lead_status',  $statuses);
        }

        // Filter by next_follow_up_date range (either start, end, or both)
        $start = $request->input('follow_up_start_date'); // e.g. '2025-08-01'
        $end   = $request->input('follow_up_end_date');   // e.g. '2025-08-31'

        if ($start) {
            $query->whereDate('next_follow_up_date', '>=', date('Y-m-d', strtotime($start)));
        }
        
        if ($end) {
            $query->whereDate('next_follow_up_date', '<=', date('Y-m-d', strtotime($end)));
        }

        // Sort + paginate (default: next_follow_up_date asc)
        $allowedSorts = ['next_follow_up_date', 'created_at', 'updated_at', 'shop_name'];
        $sort = $request->input('sort', 'next_follow_up_date');
        $dir  = strtolower($request->input('direction', 'asc')) === 'desc' ? 'desc' : 'asc';
        if (!in_array($sort, $allowedSorts, true)) {
            $sort = 'next_follow_up_date';
        }
        $query->orderBy($sort, $dir);

        $perPage = (int) $request->get('per_page', 15);
        if ($perPage <= 0) {
            $perPage = 15;
        }

        $leads = $query->paginate($perPage)->appends($request->query());

        return response()->json($leads);
    }


    /**
     * Store a new lead assigned to the logged-in sales executive.
     */
    public function store(Request $request)
    {
        try {
            $validated = $request->validate([
                'shop_name' => 'required|string|max:255',
                'contact_person' => 'required|string|max:255',
                'mobile_number' => 'required|string|max:20',
                'email' => 'nullable|email',
                'address' => 'nullable|string',
                'branches' => 'nullable|integer',
                'area_locality' => 'nullable|string',
                'pincode' => 'nullable|string',
                'gps_location' => 'nullable|string',
                'business_type' => 'nullable|integer',
                'current_system' => 'nullable|integer',
                'lead_status' => 'nullable|integer',
                'plan_interest' => 'nullable|string',
                'next_follow_up_date' => 'nullable|date',
                'meeting_notes' => 'nullable|string',
            ]);

            Log::info($validated);

            $validated['created_by'] = Auth::id();
            $validated['assigned_to'] = Auth::id();
            $validated['last_updated_by'] = Auth::id();

            $lead = Lead::create($validated);

            return response()->json([
                'success' => true,
                'data' => $lead
            ], 201);
        } catch (\Illuminate\Validation\ValidationException $e) {
            return response()->json([
                'success' => false,
                'errors' => $e->errors(),
            ], 422);
        } catch (\Exception $e) {
            return response()->json([
                'success' => false,
                'message' => 'Something went wrong',
                'error' => $e->getMessage(),
            ], 500);
        }
    }

    /**
     * Display a specific lead belonging to the logged-in sales executive.
     */
    public function show(Lead $lead)
    {
        $this->authorizeLead($lead);
        return response()->json($lead);
    }

    /**
     * Update a specific lead.
     */
    public function update(Request $request, Lead $lead)
    {
        $this->authorizeLead($lead);

        $validated = $request->validate([
            'shop_name' => 'sometimes|string|max:255',
            'contact_person' => 'sometimes|string|max:255',
            'mobile_number' => 'sometimes|string|max:20',
            'email' => 'nullable|email',
            'address' => 'nullable|string',
            'area_locality' => 'nullable|string',
            'pincode' => 'nullable|string',
            'branches' => 'nullable|integer',
            'gps_location' => 'nullable|string',
            'business_type' => 'nullable|integer',
            'current_system' => 'nullable|integer',
            'lead_status' => 'nullable|integer',
            'plan_interest' => 'nullable|string',
            'next_follow_up_date' => 'nullable|date',
            'meeting_notes' => 'nullable|string',
            'completed_at' => 'nullable|date',
        ]);

        $validated['last_updated_by'] = Auth::id();
        $lead->update($validated);

        return response()->json($lead);
    }

    /**
     * Remove a specific lead.
     */
    public function destroy(Lead $lead)
    {
        $this->authorizeLead($lead);
        $lead->delete();

        return response()->json(null, 204);
    }

    /**
     * Get leads sorted by next_follow_up_date.
     */
    public function leadsByFollowUpDate(Request $request)
    {
        $perPage = $request->get('per_page', 15); // Default 15 per page if not provided

        $noFollowUpStatuses = LeadStatus::whereIn('name', ['Sold', 'Already Using CRM', 'Not Interested', 'Using Different App'])->pluck('id');

        $totalLeads = Lead::where('assigned_to', Auth::id())
            ->whereNotIn('lead_status', $noFollowUpStatuses)
            ->count();

        $leads = Lead::where('assigned_to', Auth::id())
            ->with('leadStatusData')
            ->whereNotIn('lead_status', $noFollowUpStatuses)
            ->orderBy('next_follow_up_date', 'asc')
            ->paginate($perPage);

        return response()->json([
            'leads' => $leads,
            'total_leads' => $totalLeads
        ]);
    }

    /**
     * Ensure lead belongs to logged-in sales executive.
     */
    protected function authorizeLead(Lead $lead)
    {
        if ($lead->assigned_to !== Auth::id()) {
            abort(403, 'Unauthorized');
        }
    }
}