# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aryventory Sales Management API",
        "description": "API for managing sales executives, leads, meetings, shifts, and business operations",
        "contact": {
            "email": "support@aryventory.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Development server"
        },
        {
            "url": "https://api.aryventory.com/api",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/sales-executive/leads": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Get sales executive leads",
                "description": "Retrieve all leads assigned to the logged-in sales executive",
                "operationId": "getSalesExecutiveLeads",
                "responses": {
                    "200": {
                        "description": "Sales executive leads retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "shop_name": {
                                                "type": "string",
                                                "example": "ABC Electronics"
                                            },
                                            "contact_person": {
                                                "type": "string",
                                                "example": "John Doe"
                                            },
                                            "mobile_number": {
                                                "type": "string",
                                                "example": "9876543210"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "john@example.com"
                                            },
                                            "address": {
                                                "type": "string",
                                                "example": "123 Main Street, City"
                                            },
                                            "area_locality": {
                                                "type": "string",
                                                "example": "Downtown"
                                            },
                                            "pincode": {
                                                "type": "string",
                                                "example": "123456"
                                            },
                                            "business_type": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "current_system": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "lead_status": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "plan_interest": {
                                                "type": "string",
                                                "example": "Premium Plan"
                                            },
                                            "next_follow_up_date": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2024-01-15"
                                            },
                                            "meeting_notes": {
                                                "type": "string",
                                                "example": "Initial discussion completed"
                                            },
                                            "assigned_to": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-01-01T10:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2024-01-01T10:00:00Z"
                                            },
                                            "lead_status_data": {
                                                "type": "object",
                                                "example": {
                                                    "id": 1,
                                                    "name": "New"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sales-executive/lead-counts-by-status": {
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "Get lead counts by status",
                "description": "Get lead counts grouped by status for the logged-in sales executive",
                "operationId": "getLeadCountsByStatus",
                "responses": {
                    "200": {
                        "description": "Lead counts by status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "lead_status": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "status_name": {
                                                "type": "string",
                                                "example": "New"
                                            },
                                            "total": {
                                                "type": "integer",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": {
                "sanctum": []
            }
        }
    },
...
}